Improve comment
