+++ db_stress_tool/CMakeLists.txt
@@ -11,6 +11,7 @@ add_executable(db_stress${ARTIFACT_SUFFIX}
   db_stress_test_base.cc
   db_stress_tool.cc
   expected_state.cc
+  expected_value.cc
   multi_ops_txns_stress.cc
   no_batched_ops_stress.cc)
 target_link_libraries(db_stress${ARTIFACT_SUFFIX} ${ROCKSDB_LIB} ${THIRDPARTY_LIBS})
